Test class refactor
